Remove SpeechToText Separate model
