+++ src/base_app/models/__init__.py
@@ -1,0 +1,7 @@
+from .follower import Follower
+from .message import Message
+
+__all__ = [
+    'Follower',
+    'Message',
+]
