--- src/urls.py
@@ -1,5 +1,0 @@
-from django.urls import include, path
-
-urlpatterns = [
-    path('base/', include('base_app.urls')),
-]
