+++ src/users/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class SpeechConfig(AppConfig):
-    name = 'speech'
+    name = 'users'
