+++ src/users/models/__init__.py
@@ -1,0 +1,5 @@
+from .user import User
+
+__all__ = [
+    'User'
+]
